Update merge.yml: fix merge error
from @jasminabrar :

This seems to be occurring when there is a push during the execution of the merge action. Then when the action goes to git push the new url, the repo no longer matches the fetched repo in the action causing this error. Two potential solutions are to put ‘git pull origin main’ before all other git steps (in merge.yml) or force push using ‘git push -f origin main’.